Add -Wcpp-undef warning flag
When enabled, this new warning flag passes `-Wundef` to the C
pre-processor which causes the pre-processor to warn on uses of
the `#if` directive on undefined identifiers.

It is not currently enabled in any of the standard warning groups.

Test Plan: Make sure the two tests pass on all major platforms.

Reviewers: hvr, carter, Phyx, bgamari, austin

Reviewed By: Phyx

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D2626

GHC Trac Issues: #12752