Fix inference of partial signatures
When we had
  f :: ( _ ) => blah
we were failing to call growThetaTyVars, as we do in the
no-type-signature case, and that meant that we weren't generalising
over the right type variables.  I'm quite surprised this didn't cause
problems earlier.

Anyway Trac #12844 showed it up and this patch fixes it

(cherry picked from commit 1bfff60fc57cd564382b86bdfb1f2764ca15d44f)